wip:: --working on adaptation of project to new mathematic structures--

+++ gearoenix/render/graph/node/rnd-gr-nd-shadow-mapper.hpp
@@ -86,7 +86,7 @@ public:
     /// This will be called at the start of each frame
     void update() noexcept final;
     /// Multithreaded rendering happens in here
-    void record_shadow(const math::Mat4x4<float>& mvp, const model::Model* m, std::size_t kernel_index) noexcept;
+    void record_shadow(const math::Mat4x4<double>& mvp, const model::Model* m, std::size_t kernel_index) noexcept;
     /// This will be called at the end of each frame for pushing jobs to GPU
     void submit() noexcept final;
 };
